fix in circuit breaker

+++ cmd/lb/main.go
@@ -18,7 +18,7 @@ func main() {
 	maxFailCount := cfg.LoadBalancer.HealthCheck.MaxFailures
 	interval := cfg.LoadBalancer.HealthCheck.Interval
 
-	lb := loadbalancer.NewLoadBalancer(cfg.Backends, time.Duration(interval)*time.Second, maxFailCount, 1)
+	lb := loadbalancer.NewLoadBalancer(cfg.Backends, time.Duration(interval)*time.Second, int64(maxFailCount), 1)
 
 	log.Println("Load balancer listening on ", cfg.Server.Port)
